refactor(plugin-rows): extract plugin menu utils
